abstract move
